Se corrige renombrado de clases
